fix: make borrowed rate conversion explicit

+++ app/main.cpp
@@ -539,7 +539,9 @@ void show_statistics(const library::LibraryService& service) {
     std::cout << "borrowed_count: " << stats.borrowed_count << '\n';
 
     const double borrowed_rate =
-        stats.total_count == 0 ? 0.0 : 100.0 * stats.borrowed_count / stats.total_count;
+        stats.total_count == 0
+        ? 0.0
+        : 100.0 * static_cast<double>(stats.borrowed_count) / static_cast<double>(stats.total_count);
     std::cout << std::fixed << std::setprecision(1) << "borrowed_rate: " << borrowed_rate << "%\n";
 }
